requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
transformers==4.31.0
pillow==10.0.0
python-multipart==0.0.6
gradio==3.41.0